requirements.txt
pandas
requests
llama-index==0.11.22
llama-index-core==0.11.22
llama-index-llms-nvidia==0.2.5
python-dotenv
llama-index-readers-file==0.2.2
OpenAI==1.54.3
llama-index-vector-stores-chroma==0.3.0
llama-index-embeddings-huggingface==0.3.1
llama_index.postprocessor.nvidia_rerank==0.3.3
llama-index-readers-json==0.2.0
pyyaml
gradio==5.4.0
openpyxl